Wake on one tap, generate a roll on the next.
After waking up, it flashes " .6." until the next tap.  Then it does a
random roll (from 1 to 6) and shows that on the display, fading in and
out.  When tapped again, it rolls again.  If there is no tap for 10
seconds, it goes back to sleep.

Only simulates one die (six sided) at the moment.

Signed-off-by: Russell Steicke <russells@adelie.cx>