Simulate a die roll by flashing random numbers.
When we get a tap that means to roll, display random numbers on the
displays for a while, gradually slowing down.  At the end, when we
have the final random number, flash once briefly, then stay bright for
a few seconds, fade in and out slowly for a few more, then sleep.

A tap during the roll will restart the roll sequence.  A tap after the
final random number is displayed will start a new roll.

This is the first version that's usable as a die.

Signed-off-by: Russell Steicke <russells@adelie.cx>