Use the timer interrupts as a random source.
Our timer interrupts occur at 3906Hz, approximately.  If we have a
counter that increments on each interrupt we can grab that value at
some user generated interval and get a random 8 bit number.  We use
that value to seed one of the four bytes of the pseudo random number
generator.  3906Hz is too fast for a person to get a known value.

This prevents the pseudo random number generator from always taking
the same sequence.

Signed-off-by: Russell Steicke <russells@adelie.cx>