Always roll the die after startup.
Having a startup state which does not result in a roll is confusing,
as it's different to the state after a roll.  It was not obvious what
the thing did after it started.

So make it pause then roll after a startup tap, in the same way as it
pauses and rolls on a tap after a previous roll.

Also, remove the flashing mode from the displays, as we are no longer
using it.

Signed-off-by: Russell Steicke <russells@adelie.cx>