Add an initial design for a circuit board.
This board is designed for the 2x2 inch acrylic top and bottom case.

It would work, but is not the one getting made.  It needs to be laid
out again to put a CR2032 holder underneath, and move the vibration
sensor further inside so it is protected from prying fingers.

Signed-off-by: Russell Steicke <russells@adelie.cx>