Add a second design for a circuit board.
This one would also work, and is a better design than the first.  It
includes the battery holder soldered underneath the board.

But the circuit has been altered since work was started, and new
components are needed.  A momentary switch and a pull up resistor have
been added so a roll can be made without tapping (press the switch
instead), and another switch has been aded to disconnect the vibration
sensor.

Signed-off-by: Russell Steicke <russells@adelie.cx>