Add a double six roll mode.
It's common in games to need to roll two six sided dice.  We add that
as an extra mode, where each display will show a six roll.  The double
mode is indicated by lighting up the left side decimal point, so the
die shows "6.6" to indicate the range, and "1.2.", for example, to
show the result of a roll.

Signed-off-by: Russell Steicke <russells@adelie.cx>