Remove the use of the watchdog timer.
We're not using the watchdog for timing, and it has never been
triggered in normal (non-testing) operation, so remove it.

Also, a significant reason for removing it is that we can get into a
watchdog reset storm, where the only way to properly reset is to
remove the power.  That should be fixed, but this way is easier.

Signed-off-by: Russell Steicke <russells@adelie.cx>