Restore the use of the watchdog.
A previous commit removed the watchdog timer, in response to some
difficulties initialising it, which resulted in a watchdog reset
storm.  Removing it was not the best solution.

I have since found my error in the watchdog initialisation code.  So
restore the use of the watchdog timer, as it really should be used as
a failsafe mechanism.

Signed-off-by: Russell Steicke <russells@adelie.cx>